requirements.txt
gluoncv
PIL
numpy
torch
mxnet-cu92
